+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - (Feature) (Platform) Envoy Cache Introduction
 - (Feature) (Platform) OpenID Integration - API Extension
 - (Feature) Windows Platform CLI
+- (Feature) (Platform) Auth User Creation
 
 ## [1.2.48](https://github.com/arangodb/kube-arangodb/tree/1.2.48) (2025-05-08)
 - (Maintenance) Extend Documentation